requirements.txt
streamlit
pandas
openpyxl
requests
Pillow
opencv-contrib-python
numpy
scikit-learn
streamlit-webrtc
transformers
torch
accelerate
ultralytics
huggingface_hub
zxing-cpp
pyzbar
av
rapidocr-onnxruntime
